Fix tests invocation across CI
Employ unittest discovery CLI, fix (sometimes) broken import.
